Create Double_Integral.py
This script prompts the user to input a function of two variables, along with the lower and upper limits for both the x and y variables. The function provided by the user is then defined, and the double integral is calculated using SciPy's dblquad() function. The result of the integration is then printed out.